+++ inst/include/rxode2parse_control.h
@@ -131,6 +131,8 @@
 #define Rxc_parallel 127
 #define Rxc__zeros 128
 #define Rxc_zeroVarParamHandle 129
+#define Rxc_indLinStepSearch 130
+#define Rxc_indLinMaxIter 131
 #define RxMv_params 0
 #define RxMv_lhs 1
 #define RxMv_state 2